Feature/pundit/kevin

+++ app/controllers/properties_controller.rb
@@ -1,6 +1,11 @@
 class PropertiesController < ApplicationController
+  def new
+    @property = Property.new
+  end
+
   def create
     property = Property.new(property_params)
+    authorize property
     if property.save
       redirect_to properties_path
     else
@@ -10,11 +15,18 @@ class PropertiesController < ApplicationController
 
   def show
     @property = Property.find(params[:id])
+    authorize @property 
     @applications = @property.applications
   end
 
+  def edit 
+    @property = Property.find(params[:id])
+    authorize @property
+  end
+  
   def update
     property = Property.find(params[:id])
+    authorize property 
     if property.update(property_params)
       redirect_to properties_path
     else
@@ -24,6 +36,7 @@ class PropertiesController < ApplicationController
 
   def destroy
     property = Property.find(params[:id])
+    authorize property
     if property.destroy
       redirect_to properties_path
     else
